# PricingManagement/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Management API",
        "description": "API for managing payment transactions and their status",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create Payment",
                "description": "Initiates a new payment transaction for an order.",
                "operationId": "82f805bcf2dd8149b522bead8ae17044",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "orderId",
                                    "amount",
                                    "userId"
                                ],
                                "properties": {
                                    "orderId": {
                                        "type": "string",
                                        "example": "order123"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "userId": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created"
                    },
                    "500": {
                        "description": "Application error"
                    }
                }
            }
        },
        "/api/payment/{paymentId}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get Payment Status",
                "description": "Retrieves the status of a specific payment.",
                "operationId": "d8ee8d8650c0c6febca100a071230af2",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "The ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "paymentId": {
                                            "type": "string",
                                            "example": "payment_123"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 25000
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-07-21T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Application error"
                    }
                }
            }
        },
        "/api/payments/user/{userId}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "List All Payments for a User",
                "description": "Retrieves all payments made by a specific customer.",
                "operationId": "395f2548285b4168046f6b36c0e76397",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "paymentId": {
                                                "type": "string",
                                                "example": "payment_123"
                                            },
                                            "orderId": {
                                                "type": "string",
                                                "example": "order123"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "completed"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 25000
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-07-21T10:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Application error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Payment",
            "description": "Payment"
        }
    ]
}